requirements.txt
monarchmoney==0.1.15
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.3
openpyxl==3.1.5
SQLAlchemy==2.0.36
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

